created new calculatermain file - home to calculatermain
