+++ calculatorapp/src/App.js
@@ -11,6 +11,7 @@ export default function App() {
    <Routes>
   
     <Route path="/" element={<Home/>}/>
+  
    </Routes>
    </BrowserRouter>
   );
